chore(staticGenerator): modulize static generator, generate author post list json file
